keyword_issue

+++ src/api.h
@@ -10,4 +10,4 @@ void fetch_api_with_url(const char* url, char* buffer, int buffer_size);
 void fetch_api(char* buffer, int buffer_size);
 
 // collect quotes by keywords
-void fetch_quote_by_kw(char* buffer, int buffer_size, const char* kw);
+void fetch_quote_by_kw(char* buffer, size_t buffer_size, const char* kw);
